fix(#2070): smooth game switching in directories without page reload
Use state-based game tracking and window.history.replaceState to swap
games in-place when clicking from the Directories tab, avoiding the
Next.js App Router remount that caused a flash on every click.
Shift-click still opens in a new tab as before.
